+++ src/outputs/discord.py
@@ -13,6 +13,7 @@ import requests
 from config import config
 from formatters.discord import DiscordFormatter
 from formatters.discord_grafana import GrafanaDiscordFormatter
+from formatters.discord_portainer import PortainerDiscordFormatter
 from formatters.discord_qnap import QNAPDiscordFormatter
 from formatters.discord_truenas import TrueNASDiscordFormatter
 from formatters.discord_unifi import (
@@ -33,6 +34,7 @@ class DiscordOutput:
 
         self.source_formatters = {
             "grafana": GrafanaDiscordFormatter(),
+            "portainer": PortainerDiscordFormatter(),
             "qnap": QNAPDiscordFormatter(),
             "truenas": TrueNASDiscordFormatter(),
             "unifi_drive": UniFiDriveDiscordFormatter(),